5_完善删除:删除前检查是否包含检查组关联 增加全局异常类 分页查询使用pageHelper分页

+++ health_parent/health_service/src/main/java/com/itheima/health/service/impl/CheckItemServiceImpl.java
@@ -1,13 +1,15 @@
 package com.itheima.health.service.impl;
 
 import com.alibaba.dubbo.config.annotation.Service;
+import com.github.pagehelper.Page;
+import com.github.pagehelper.PageHelper;
 import com.itheima.health.dao.CheckItemDao;
 import com.itheima.health.entity.PageResult;
 import com.itheima.health.entity.QueryPageBean;
+import com.itheima.health.exception.HealthException;
 import com.itheima.health.pojo.CheckItem;
 import com.itheima.health.service.CheckItemService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -30,22 +32,42 @@ public class CheckItemServiceImpl implements CheckItemService {
 
     @Override
     public PageResult<CheckItem> findPage(QueryPageBean queryPageBean) {
-
+        //使用原本的方式查询
+        /*
         //计算从哪开始查
         Integer currentPage = queryPageBean.getCurrentPage();
         queryPageBean.setCurrentPage((currentPage-1)*queryPageBean.getPageSize());
-
         //查询结果
         List<CheckItem> pageRows = checkItemDao.findPage(queryPageBean);
-
         //查询条数
         Long count=checkItemDao.getCount(queryPageBean);
+        */
 
-        return new PageResult<CheckItem>(count,pageRows);
+        //使用pageHelper插件
+        PageHelper.startPage(queryPageBean.getCurrentPage(), queryPageBean.getPageSize());
+
+        //判断是否有条件
+        String queryString = queryPageBean.getQueryString();
+        if(queryString!=null && queryString.length()>0){
+            queryPageBean.setQueryString("%"+queryString+"%");
+        }
+
+        //获取查询结果  ,拿到条件下所有的结果
+        //使用pageHelper分页
+        Page<CheckItem> page = checkItemDao.findPage(queryPageBean);
+
+        //能拿到所有的条件结果就 等于 拿到了所有的结果的条数 , 就不用再执行一次获取 数量的sql语句
+        //从page拿到处理的结果
+        return new PageResult<CheckItem>(page.getTotal(),page.getResult());
     }
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Integer id) throws HealthException{
+        int countByCheckItemId = checkItemDao.findCountByCheckItemId(id);
+
+        if(countByCheckItemId>0){
+            throw new HealthException("这个检查项有关联的检查组,不能删除");
+        }
         checkItemDao.deleteById(id);
     }
 }
